dagster_project/pyproject.toml
[project]
name = "dagster_extract_transform_duckdb"
requires-python = ">=3.9,<=3.13.3"
version = "0.1.0"
dependencies = [
    "dagster==1.11.2",
    "datetime>=5.5",
    "dbt-core>=1.10.4",
    "dbt-duckdb>=1.9.4",
    "duckdb>=1.3.2",
    "numpy>=2.0.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "shopifyapi>=12.7.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_extract_transform_duckdb"
registry_modules = [
    "dagster_extract_transform_duckdb.components.*",
]


